Supprot inplace + vectorization
