+++ backend-comparison/benches/binary.rs
@@ -23,10 +23,9 @@ impl<B: Backend, const D: usize> Benchmark for BinaryBenchmark<B, D> {
 
     fn execute(&self, (lhs, rhs): Self::Args) {
         // Choice of add is arbitrary
-        let mut tmp = lhs.clone();
-        for _ in 0..10 {
-            tmp = tmp.clone() + rhs.clone() / 3.0;
-        }
+        let a = Tensor::<B, D>::random(self.shape.clone(), Distribution::Default, &self.device);
+        let b = Tensor::<B, D>::random(self.shape.clone(), Distribution::Default, &self.device);
+        let out = a + b;
         //  B::float_add(
         //      lhs.clone().into_primitive().tensor(),
         //      rhs.clone().into_primitive().tensor(),
