+++ crates/burn-jit/src/fusion/on_write/mod.rs
@@ -4,3 +4,4 @@ pub(crate) mod ir;
 pub(crate) mod kernel;
 
 pub mod trace;
+pub(crate) mod trace_builder;
